Makefile: fix verilator incantations
Verilator seems to have a new dependency (verilated_threads.cpp) on
itself. Since we're not using the verilator generated makefile for
everything this was breaking us. Let's add it as a dependency
(after we massage some of our rules).

Also used /* verilator public */ in a couple of places as it seems
to be more reliable.

TEST=make clean tests
No more "verilated.cpp:(.text+0xdbe6): undefined reference to `VlThreadPool::VlThreadPool(VerilatedContext*, unsigned int)'" error.